Able to only react to changes to files under VC now (with contents.lc.yaml as a special case). Code is sloppy, but can be demoed. The envsubst plugin should actually be able to run regardless of whether cluster can be loaded (because it can no longer run after an error is introduced into contents.lc.yaml). May need a fourth plugin phase for that, possibly even one which is specified in a different file. That would also remedy the issue with auto-saving editors like VSC making the system too trigger happy.

+++ rust-workspace/envsubst_pre_contents_plugin/src/lib.rs
@@ -57,10 +57,22 @@ impl ClusterProcessingPlugin for EnvsubstPreContentsPlugin {
         cluster_path: &Path,
         _cluster: &domain::Cluster,
     ) -> Result<HashSet<ArtifactMapping>, anyhow::Error> {
+        println!("Running envsubst plugin.");
         let input_file_contents =
             std::fs::read_to_string(cluster_path.join("pre-contents.lc.yaml"))?;
-        let replacement_text = substitute(input_file_contents, &std::env::vars().filter(|e| e.0 == "APPLICATION_SOURCE_DIR").collect())?;
-        std::fs::write(cluster_path.join("contents.lc.yaml"), replacement_text)?;
+        println!("Read input file.");
+        let replacement_file_contents =
+            std::fs::read_to_string(cluster_path.join("contents.lc.yaml"))?;
+        let replacement_text = substitute(
+            input_file_contents,
+            &std::env::vars()
+                .filter(|e| e.0 == "APPLICATION_SOURCE_DIR")
+                .collect(),
+        )?;
+        if replacement_text != replacement_file_contents {
+            println!("Rewriting file.");
+            std::fs::write(cluster_path.join("contents.lc.yaml"), replacement_text)?;
+        }
         Ok(HashSet::new())
     }
 }
